+++ src/main/java/net/sf/jabref/gui/importer/fetcher/GeneralFetcher.java
@@ -77,12 +77,6 @@ public class GeneralFetcher extends SidePaneComponent implements ActionListener 
         String[] choices = new String[fetcherArray.length];
         for (int i = 0; i < fetcherArray.length; i++) {
             choices[i] = fetcherArray[i].getTitle();
-            //choices[i] = new JLabel(fetchers.get(i).getTitle(), new ImageIcon(fetchers.get(i).getIcon()),
-            //        JLabel.HORIZONTAL);
-            /*if (fetchers.get(i).getOptionsPanel() != null)
-                optionsPanel.add(fetchers.get(i).getOptionsPanel(), String.valueOf(i));
-            else
-                optionsPanel.add(new JPanel(), String.valueOf(i));*/
         }
         JComboBox<String> fetcherChoice = new JComboBox<>(choices);
         int defaultFetcher = Globals.prefs.getInt(JabRefPreferences.SELECTED_FETCHER_INDEX);
@@ -97,26 +91,7 @@ public class GeneralFetcher extends SidePaneComponent implements ActionListener 
         HelpAction help = new HelpAction(activeFetcher.getHelpPage());
         JButton helpBut = help.getHelpButton();
         helpBut.setEnabled(activeFetcher.getHelpPage() != null);
-
-        //optionsCards.show(optionsPanel, String.valueOf(defaultFetcher));
-
-        /*fetcherChoice.setRenderer(new ListCellRenderer() {
-            JLabel label = new JLabel();
-            public Component getListCellRendererComponent(JList jList, Object o, int i, boolean isSelected,
-                boolean cellHasFocus) {
-                JLabel theLab = (JLabel)o;
-                label.setIcon(theLab.getIcon());
-                label.setText(theLab.getText());
-                if (cellHasFocus) {
-                    label.setBackground(UIManager.getDefaults().getColor("ComboBox.selectionBackground").darker());
-                    label.setForeground(UIManager.getDefaults().getColor("ComboBox.foreground"));
-                } else {
-                    label.setBackground(UIManager.getDefaults().getColor("ComboBox.background"));
-                    label.setForeground(UIManager.getDefaults().getColor("ComboBox.foreground"));
-                }
-                return label;
-            }
-        });*/
+        
         fetcherChoice.addActionListener(actionEvent -> {
             activeFetcher = fetcherArray[fetcherChoice.getSelectedIndex()];
             Globals.prefs.putInt(JabRefPreferences.SELECTED_FETCHER_INDEX, fetcherChoice.getSelectedIndex());