+++ src/main/java/net/sf/jabref/logic/util/FileExtensions.java
@@ -12,7 +12,7 @@ import net.sf.jabref.logic.l10n.Localization;
 public enum FileExtensions {
     //important: No dot before the extension!
     BIBTEX_DB(String.format("%1s %2s", "BibTex", Localization.lang("Database")), "bib"),
-    BIBTEXML(Localization.lang("%0 file", "BibTeXML"), "bibx"),
+    BIBTEXML(Localization.lang("%0 file", "BibTeXML"), "bibx", "xml"),
     BILBIOSCAPE(Localization.lang("%0 file", "Biblioscape"), "txt"),
     COPAC(Localization.lang("%0 file", "Copac"), "txt"),
     ENDNOTE(Localization.lang("%0 file", "Endnote/Refer"), "ref", "enw"),